Repo org: per-section folders with index.tsx, move section-specific UI in
- Created folders: sections/{about,contact,experience,landing,services,skills}
- Moved section components to index.tsx in each folder
- Moved InteractiveElements to landing/, JobCard & JobView to experience/, ServiceCard to services/
- Updated imports across app and sections
- Clean separation: shared UI in ui/, section-specific UI colocated with sections
