+++ src/components/sections/experience/index.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react'
-import PageSection from '../ui/PageSection'
-import JobCard from '../ui/JobCard'
-import JobView from '../JobView'
+import PageSection from '../../ui/PageSection'
+import JobCard from './JobCard'
+import JobView from './JobView'
 import fdaLogo from '../../assets/fda-logo.png'
 import softhreadLogo from '../../assets/Softhread_Logo.jpg'
 import delphiLogo from '../../assets/DELPHI-DIGITAL-MASTER-LOGO.jpg'
