+++ src/components/sections/landing/InteractiveElements.tsx
